Initial attempt at creating Pong with raylib
The following is just an initial attempt at getting raylib up and running to see what it is like to program using raylib. The game is simple and semi-functional right now. However, there are a slew of TODOs to be fixed later on.

For now, we'll push the current state of the game as-is.